python/ai_how/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-how"
version = "0.1.0"
description = "AI-HOW: Python CLI orchestrator for managing the Hyperscaler on Workstation clusters"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "AI-HOW Team", email = "devnull@example.com" }
]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "jsonschema>=4.21.0",
  "PyYAML>=6.0",
  "libvirt-python>=9.0.0",
  "jinja2>=3.1.0",
  "expandvars>=1.1.2"
]

[project.scripts]
ai-how = "ai_how.cli:app"

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "pre-commit>=3.0.0",
  "mypy>=1.0.0",
  "pyright>=1.1.0",
  "ruff>=0.1.0",
  "black>=23.0.0",
  "types-PyYAML>=6.0.0",
  "types-jsonschema>=4.0.0",
  "nox>=2023.4.22",
  "nox-uv>=0.6.0",
]
docs = [
  "mkdocs>=1.5.0",
  "mkdocs-material>=9.0.0",
  "mkdocstrings[python]>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_how"]

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
  "/docs",
  "/README.md",
  "/pyproject.toml",
  "/LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W", "C4", "UP", "ARG", "SIM", "TCH", "TID", "Q"]
ignore = ["UP007"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
strict_optional = true
no_implicit_optional = true
pretty = true
show_error_context = true
namespace_packages = true
explicit_package_bases = true
packages = ["ai_how"]
mypy_path = "stubs"
exclude = [
    "^build/",
    "^.nox/",
    "^.venv/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "--strict-markers",
  "--strict-config",
  "--cov=src/ai_how",
  "--cov-report=term-missing",
  "--cov-report=html",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pyright>=1.1.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
]

[tool.pyright]
include = ["src"]
exclude = [
    "**/__pycache__",
    ".nox",
    ".venv",
    "build",
    "dist",
]
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.11"
typeCheckingMode = "basic"
# Stricter checks to catch issues like missing keyword arguments
reportCallIssue = "error"
reportArgumentType = "error"
reportGeneralTypeIssues = "error"
# Helpful for catching dynamic attribute access issues
reportOptionalMemberAccess = "warning"
reportOptionalSubscript = "warning"
